move collision group consts to shared lib
